Implement the option "diffn --ifdefs".
The command "./diffn --ifdefs foo bar" will generate a merge that's
similar to "./diffn foo bar", but using embedded #ifdef/#endif lines
instead of prefixed a/b/c/etc. Furthermore, we'll play tricks with
blank lines in order to make the #ifdefs appear more human-friendly.
For example, where a plain (patience) diff would produce

     void foo() {
         puts("hello");
     }

    +void addendum() {
    +    puts("cruel");
    +}
    +
     void bar() {
         puts("world");
     }

we'll actually produce the following lossy-but-nice output:

     void foo() {
         puts("hello");
     }

     #if defined(V2)
     void addendum() {
         puts("cruel");
     }
     #endif /* V2 */

     void bar() {
         puts("world");
     }

(Notice the positioning of the blank lines relative to the
preprocessor directives. This is not a mistake.)

This is the main functionality I've been aiming for. The only
remaining changes will be to improve the quality of the merge
and allow the command-line user to customize the #define macros
used in the #ifdef lines. Right now we just use V1, V2, etc.