Add a tests/ directory with some simple tests.
These tests all pass. One test that doesn't yet pass is not
checked in: that's the problematic merge pointed out on
http://bramcohen.livejournal.com/73318.html#t1081702
Specifically, diff(ZAXCX, AXBXCX) will begin by matching the
prefix "" and the suffix "XCX", which is ugly; we really want
to match the suffix "CX" so that we can match "AX" later.
I have an idea to fix this, but have not yet implemented it.