Add the -D option.
The diffn utility now supports a "-D" option similar to that
provided by GNU diff. The difference, of course, is that diffn
can take (exactly) as many "-D" options as there are files on
the command line. The user-supplied names replace the default
"V1", "V2", etc. in the merged file.