Add tests for simple nested #ifs.
"nested-append.sh" demonstrates nice behavior that just happens
accidentally. "nested-prepend.sh" demonstrates ugly behavior,
which would be easy to fix in this particular case --- just scan
forward until you find a mask that's not a subset of the new mask,
and if that mask is in fact a *super*set of the new mask, then
output the #if for the superset (and push it on the stack) before
the current #if. But I'm not sure that the simple fix will really
produce objectively better results, so I'm leaving "nested-append.sh"
alone for now. It still verifies that we don't regress or crash on
this input, which is a good thing.