Add a warning for input files already containing #ifdefs.
Also, add errors if the input files' #ifdefs are unbalanced;
in that case I don't think "diffn --ifdefs" can possibly produce
meaningful output.

I have some ideas about how to handle #ifdefs in the input files,
but I have not yet implemented them, so for now I'm just making
diffn produce a warning about possibly bogus output. A simple
example would be:

    ==File 1==  ==File 2==  ==Merged output==
    #if A       #if A       #if A
      foo         foo         foo
    #endif A    #else A     #if defined(V2)
                  bar       #else A
                #endif A      bar
                            #endif /* V2 */
                            #endif A

Perniciously, the merged output is a legal nesting, but its
meaning is NOT the same as the meaning of the correct merge:

    #if defined(V1)
    #if A
      foo
    #endif
    #endif  /* V1 */
    #if defined(V2)
    #if A
      foo
    #else
      bar
    #endif
    #endif  /* V2 */

Ideally, we'd produce the following very clever merge. We certainly
have enough information to verify that this merge would be correct;
the hard part is actually generating it.

    #if A
      foo
    #else
    #if defined(V2)
      bar
    #endif  /* V2 */
    #endif