Correctly (if naively) handle inputs already containing #if/#endif.
After this commit, we handle the problematic merge posed in my last
commit message:

        ==File 1==  ==File 2==  ==Old output==
        #if A       #if A       #if A
          foo         foo         foo
        #endif A    #else A     #if defined(FOO)
                      bar       #else A
                    #endif A      bar
                                #endif /* FOO */
                                #endif A
        ==New output==
        #if defined(NOFOO)
        #if A
          foo
        #endif A
        #endif /* NOFOO */
        #if defined(FOO)
        #if A
          foo
        #else A
          bar
        #endif A
        #endif /* FOO */

Notice that the old merge produces a valid nesting of pp-directives, but
it has a very different meaning from the correct merge. Also, notice that
the old merge is equivalent to that produced by GNU's "diff -D".

Admittedly the new merge is about twice as long as it really needs to be
in this example, but smartening it up is going to be a little tricky.

Roadmap of high-priority issues:

    - Correctly handle backslash-escapes.
    - Produce output containing "#if !defined", "#elif", and "#else".
    - Allow using as a drop-in replacement for "diff -D" by passing
      one fewer -D argument than there are files on the command line.
    - Warn if some -D macro is already in use within the file.