Improve handling of nested #ifs in input; add regression tests.
The regression test pp-7-AA.sh looks like this:

    ==File A==  ==File B==  ==Merged output==
    #if A       #if A       #if A
                            #if defined(V1)
    #if B                   #if B
    #endif B                #endif B
                            #endif /* V1 */
    #endif A    #endif A    #endif A

My last revision didn't attempt this clever merge, but this
new revision does. We do have to watch out for this next example,
though; see pp-6*.sh.

    ==File A==  ==File B==
                #if foo
    match1      match1
    #if bar
                #endif foo
    match2      match2
    #endif bar

Here, the first #endif in File B does not terminate the #if-chain
begun by "#if foo", because there is still a nest open in File A.
Notice that the diff would arguably look prettier if we swapped
the order of "#if bar" and "#endif foo", but I don't yet see any
objective way of deciding that.

I've added several new tests.