Add --unified diff mode (two files only), plus tests.
Also add the --help option, and do some minor cleanup.

The tests for --unified mode are pretty minimal, but
they do test against the actual system "diff" utility
(we assume it's GNU diff) and make sure the output
of "./diffn -U<num>" is byte-for-byte identical to
that of "diff -U<num>".

I'm aware that my handling of single-letter options
and -u<num> doesn't quite match GNU's yet, but I'm
not sure I want to add a dependency on GNU getopt.