Initial implementation of the -r (--recursive) option.
This tree-walking code is really icky and probably fragile.
It's also not standard C++ (although it *is* POSIX, mostly).
I've tried to segregate it into its own source file so that
it can be snipped out by anyone trying to compile diffn on
a non-POSIX system.

    diffn -r -DA -DB DirectoryA DirectoryB -o DirectoryC

The above command line will recursively walk down DirectoryA
and DirectoryB, and write the merged output to DirectoryC.
It's equivalent to running

    mkdir DirectoryC
    diffn -r -DA -DB Directory{A,B}/one.c -o DirectoryC/one.c
    diffn -r -DA -DB Directory{A,B}/two.c -o DirectoryC/two.c
    mkdir DirectoryC/subdir
    diffn -r -DA -DB Directory{A,B}/subdir/three.c -o DirectoryC/subdir/three.c

and so on. If a file exists in A but not in B (or vice versa),
the missing file is treated as if it were zero-length.