Fix a pretty bad bug in my premature optimizations.
My clever optimization to query the StringSet for unique lines
instead of doing a linear search doesn't actually work. The
problem is that a line "foo" may be unique in version A and
unique in version B, but not unique (because not present) in
the *particular subrange* of B that we happen to be looking at
in this recursive invocation of add_vec_to_diff().

Vice versa, "foo" may be duplicated in B, but unique in the
particular subrange of B we're looking at.

An embarrassingly simple test case exposes this bug:

    ==File A==    ==File B==
    foo           bar
    bar           foo

Added as "tests/regress-1.sh".