Avoid fdopendir() for non-POSIX (Mac OS X) systems.
On OS X, the best we can do is to use opendir() on the directory's
name again. This raises time-of-check-to-time-of-use issues, but
I don't think there's anything we can do about it. (And it's not
a *security* problem; the worst that can happen is that diffn
will produce bogus output or crash. Which it might do anyway, if
you're in the habit of moving subdirectories around or rewriting
them while diffn is in the middle of a readdir() loop.)