Coalesce #endif directives whose differences aren't significant.
A common anti-pattern I've seen in real code is

    ab#if SOMETHING
    ab  foo
    a #endif // foo
     b  bar
     b#endif // foo bar
    ab  baz

In this case, the only possible merge is to duplicate the entire range,
even though the only difference between A and B is in a non-significant
part of the #endif directive.
To deal with this anti-pattern, we'll look for adjacent but mutually
exclusive blocks-ending-in-#endif and merge them, yielding

    ab#if SOMETHING
    ab  foo
     b  bar
    ab#endif // foo bar
    ab  baz

We'll just arbitrarily pick the last file's #endif line. We should
probably pick the most popular line instead, but that would require
more code.

Notice that this is a stupid heuristic. It won't catch

    ab#if SOMETHING
    ab  foo
    a #endif // foo
     b#endif // bar
     b  baz

However, if there's a shared line (e.g., a blank line) following
the #endifs, the heuristic will usually work all right.