Deal with multi-line constructs (comments, literals...)
This had been on my to-do list for a while. As illustrated
by the two new tests, we can now handle cases like this:

    ==File A==    ==File B==    ==Naive Merge==
    /* hello      /* hello      /* hello
       world */      sailor */  #if defined(A)
                                   world */
                                #endif /* A */
                                #if defined(B)
                                   sailor */
                                #endif /* B */

The new code (which is surprisingly short!) reuses the existing
logic for splitting up #if-#elif-#endif ranges. The new
"CStateMachine" class is a bit of a hack; I don't like how
much code is still duplicated between verify.cc and ifdefs.cc.