Try out a new algorithm for assigning #if/#endif locations.
This algorithm, which I've called use_synchronization_points(),
is a little easier to explain and almost trivial to prove correct.
Unfortunately, it is strictly less useful than the old one.

Basically, the rule is "insert #if/#endif only at sync points",
where a sync point is defined as any linebreak that isn't inside
a multiline construct and also isn't inside any #if construct.
It's the latter condition that makes this algorithm basically
useless (because it can't deal sensibly with header guards), but
I didn't realize it until I'd already written the code. So here
it is.

The following regression tests currently fail in this branch:

    coalesce-endifs.sh
    header-guard.sh
    pp-5.sh
    pp-6a.sh
    pp-7-AA.sh