Implement more-GNUish "#ifdef", and the -DNAME=VALUE option.
Old behavior: --ifdef=A would give you conditionals on defined(A),
and there was no way to get an #if on an arbitrary condition.

New behavior: --ifdef=A gives you conditionals on defined(A), but
we special-case "#if defined(A)" down to "#ifdef A".  To get
perfect control over the text of the condition, use --if=A.

As potentially useful special cases, we'll handle -DNAME as a
synonym for --ifdef=NAME and -DNAME=VALUE as a synonym for
--if=NAME==VALUE. Notice that the latter sacrifices robustness
for readability; if you're going to use arbitrary VALUEs, you
should use "--if=(NAME==(VALUE))" instead.