Coalesce identical vectors in simply_concatenate().
Okay, so simply_concatenate() wasn't quite the right name
for this function. If we have a diff like this:

    a  dfirst version
     bcdsecond version

then what we actually want to see is not

    a   first version
     b  second version
      c second version
       dfirst version
       dsecond version

but rather

    a   first version
     bc second version
       dfirst version
       dsecond version

This requires comparing whole vectors for equality, which
sounds expensive, but in practice simply_concatenate() is
called only on ranges that differ in odd ways, and those
ranges are likely to be short. I do need to write some
"pathologically large input" tests, though.