Add #elif/#else support; fix bugs in CStateMachine.
We continue to use #ifdef/#endif by default (or if the user
explicitly passes the --simple option), but if you invoke
"diffn --complex", we'll use the full complement of

    #if A
      ...
    #elif B
      ...
    #else /* C || D */
      ...
    #endif /* A || B || C || D */

The "#else" is commented with the versions it covers, and
the "#endif" is commented with all of the versions covered
by the entire construct. One special case is implemented for
convenience: If you pass -DFOO=A -DFOO=B etc., and an #if/#else
construct covers all N versions, then the #endif of that
construct will be commented with /* FOO */.

    #if FOO==A
      ...
    #else /* FOO==B */
      ...
    #endif /* FOO */

Also mixed in with this revision are some major improvements
to "state-machine.cc", and regression tests therefor.