Hack in a -t option to normalize whitespace.
When you pass the -t option, the input system will filter each
input file by removing trailing whitespace and expanding spaces
to tabs. This happens before any diffing, so "difdef -t -DA a.txt"
will produce output that is not identical to "cat a.txt". It is
similar to GNU diff's -t/--expand-tabs option (hence the name),
but not intended to be a drop-in replacement.

In fact I hope this is just a temporary hack. A better solution
would roll both this and coalesce_endifs() into a general solution:
instead of doing string comparisons, we should put each line
into an "equivalence class" of similar lines, diff based on the
equivalence classes, and then take a vote among all the
versions' corresponding lines to decide which one of the
insignificantly-different lines to print in the merged output.
But that will take quite some refactoring; we'll have to stop
passing around std::vector<const std::string *> everywhere.