Fix a major bug in the brace-heuristic logic.
Revision 6e0ca9bb contained a bug. We did the correct thing when
presented with a case like threehalves.sh, which contains an "ab"
section followed by an "a" section followed by another "ab" section.
In that case, it's okay to swap "ab" lines from one end of the diff
to the other as long as they compare equal; conceptually we're
choosing where to insert the new code.

But when we have a case like regress-2.sh --- an "a" section
followed by an "ab" section followed by another "a" section ---
swapping identical lines from one end of the diff to the other
actually reorders lines within "a"!

I'm actually not 100% confident in this bugfixed algorithm either,
but at least it passes regress-2.sh.