Regression test for max number of open file handles.
This is a regression test for the bug fixed in revision e831e283.
It creates three directories, each with 10 subdirectories of 100 files
apiece; and then with 100 subdirectories of 10 files apiece. difdef
should be able to handle either of these scenarios without exceeding
256 open file handles --- although its procedure for dealing with
file handles certainly isn't as economical as it could be. Ideally
we'd be able to diff any N directory trees, each K levels deep, with
only N+K file handles. Right now it's more like N*K.